% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{property_boolean}
\alias{property_boolean}
\title{Build a boolean property descriptor}
\usage{
property_boolean(description = NULL, default = NULL, required = FALSE)
}
\arguments{
\item{description}{Optional description.}

\item{default}{Optional default value.}

\item{required}{Whether this property is required.}
}
\value{
A property descriptor list.
}
\description{
Build a boolean property descriptor
}
\examples{
property_boolean("Verbose")
}
