% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{remove_resource}
\alias{remove_resource}
\title{Remove a resource by URI}
\usage{
remove_resource(mcp, uri)
}
\arguments{
\item{mcp}{An \code{McpServer}.}

\item{uri}{Resource URI.}
}
\value{
\code{mcp}, invisibly.
}
\description{
Remove a resource by URI
}
